move /internal/git to /pkg/git
