+++ commands/ci/lint/lint.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/profclems/glab/api"
 	"github.com/profclems/glab/commands/cmdutils"
-	"github.com/profclems/glab/internal/git"
+	"github.com/profclems/glab/pkg/git"
 
 	"github.com/MakeNowJust/heredoc"
 	"github.com/spf13/cobra"
