+++ commands/ci/trace/trace.go
@@ -13,7 +13,7 @@ import (
 	"github.com/profclems/glab/api"
 	"github.com/profclems/glab/commands/ci/ciutils"
 	"github.com/profclems/glab/commands/cmdutils"
-	"github.com/profclems/glab/internal/git"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/pkg/utils"
 
 	"github.com/AlecAivazis/survey/v2"
