+++ commands/ci/view/view.go
@@ -14,7 +14,7 @@ import (
 	"github.com/profclems/glab/api"
 	"github.com/profclems/glab/commands/ci/ciutils"
 	"github.com/profclems/glab/commands/cmdutils"
-	"github.com/profclems/glab/internal/git"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/pkg/utils"
 
 	"github.com/MakeNowJust/heredoc"
