+++ commands/cmdutils/cmdutils.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/profclems/glab/internal/config"
 
-	"github.com/profclems/glab/internal/git"
+	"github.com/profclems/glab/pkg/git"
 )
 
 const (
