+++ commands/cmdutils/cmdutils_test.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/acarl005/stripansi"
 	"github.com/profclems/glab/api"
-	"github.com/profclems/glab/internal/git"
 	"github.com/profclems/glab/internal/glrepo"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/pkg/prompt"
 	"github.com/stretchr/testify/assert"
 	"github.com/xanzy/go-gitlab"
