+++ commands/cmdutils/factory.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/profclems/glab/api"
 	"github.com/profclems/glab/internal/config"
-	"github.com/profclems/glab/internal/git"
 	"github.com/profclems/glab/internal/glinstance"
 	"github.com/profclems/glab/internal/glrepo"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/pkg/iostreams"
 	"github.com/xanzy/go-gitlab"
 )
