+++ commands/issue/note/issue_note_create_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/profclems/glab/api"
 	"github.com/profclems/glab/commands/cmdutils"
 	"github.com/profclems/glab/internal/config"
-	"github.com/profclems/glab/internal/git"
 	"github.com/profclems/glab/internal/glrepo"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/pkg/httpmock"
 	"github.com/profclems/glab/pkg/prompt"
 	"github.com/profclems/glab/test"
