+++ commands/mr/checkout/mr_checkout.go
@@ -9,7 +9,7 @@ import (
 	"github.com/profclems/glab/api"
 	"github.com/profclems/glab/commands/cmdutils"
 	"github.com/profclems/glab/commands/mr/mrutils"
-	"github.com/profclems/glab/internal/git"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/tcnksm/go-gitconfig"
 
 	"github.com/spf13/cobra"
