+++ commands/mr/create/mr_create_test.go
@@ -19,8 +19,8 @@ import (
 	"github.com/profclems/glab/api"
 	"github.com/profclems/glab/commands/cmdutils"
 	"github.com/profclems/glab/internal/config"
-	"github.com/profclems/glab/internal/git"
 	"github.com/profclems/glab/internal/glrepo"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/pkg/httpmock"
 	"github.com/profclems/glab/test"
 	"github.com/stretchr/testify/assert"
