+++ commands/mr/diff/diff_test.go
@@ -17,8 +17,8 @@ import (
 	"github.com/jarcoal/httpmock"
 	"github.com/profclems/glab/commands/cmdutils"
 	"github.com/profclems/glab/internal/config"
-	"github.com/profclems/glab/internal/git"
 	"github.com/profclems/glab/internal/glrepo"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
