+++ commands/mr/note/mr_note_create_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/profclems/glab/commands/cmdtest"
 	"github.com/profclems/glab/commands/cmdutils"
 	"github.com/profclems/glab/internal/config"
-	"github.com/profclems/glab/internal/git"
 	"github.com/profclems/glab/internal/glrepo"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/pkg/httpmock"
 	"github.com/profclems/glab/pkg/prompt"
 	"github.com/profclems/glab/test"
