+++ commands/project/clone/repo_clone.go
@@ -11,9 +11,9 @@ import (
 	"github.com/profclems/glab/api"
 	"github.com/profclems/glab/commands/cmdutils"
 	"github.com/profclems/glab/internal/config"
-	"github.com/profclems/glab/internal/git"
 	"github.com/profclems/glab/internal/glinstance"
 	"github.com/profclems/glab/internal/glrepo"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/xanzy/go-gitlab"
