+++ internal/glrepo/resolver.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/profclems/glab/api"
-	"github.com/profclems/glab/internal/git"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/pkg/prompt"
 
 	"github.com/xanzy/go-gitlab"
