+++ internal/glrepo/resolver_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/profclems/glab/api"
-	"github.com/profclems/glab/internal/git"
+	"github.com/profclems/glab/pkg/git"
 	"github.com/profclems/glab/pkg/prompt"
 	"github.com/stretchr/testify/assert"
 	"github.com/xanzy/go-gitlab"
